EditButton bug fixed!

+++ app/src/main/java/com/example/proje/UI/CarListFragment.java
@@ -90,7 +90,7 @@ public class CarListFragment extends Fragment {
         args.putString("car_brand", car.getBrand());
         args.putString("car_model", car.getModel());
         args.putInt("car_year", car.getYear());
-        args.putDouble("car_price", car.getPrice());
+        args.putFloat("car_price", (float) car.getPrice());
         args.putString("car_userId", car.getUserId());
         
         Navigation.findNavController(requireView())
